requirements.txt
gymnasium
numpy
tensorflow
tensorflow-probability
distrax
flax
ml_collections
tqdm
chex
optax
absl-py
scipy
wandb
einops
imageio
moviepy
pre-commit == 3.3.3
ipdb
seaborn
